+++ CHANGELOG.md
@@ -16,6 +16,7 @@
  - Added `Control-Y` readline binding.
  - Added ability to control whether links from secondary instances open in a new window.
  - Added `luakit.resource_path` property to control where luakit searches for resource files.
+ - Added `lousy.util.find_resource` function.
 
 ### Changed